Upgraded Ruby, Rubocop, Changed Top Level API, README
