+++ .rubocop.yml
@@ -2,3 +2,9 @@ inherit_from: .rubocop_todo.yml
 
 Metrics/LineLength:
   Max: 100
+
+Metrics/BlockLength:
+  ExcludedMethods: ['it', 'describe', 'context', 'let', 'specify']
+
+AllCops:
+  TargetRubyVersion: 2.3
