+++ .ruby-version
@@ -1,1 +1,1 @@
-2.3.7
+2.6.0
