+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+# 2.0.0 (January 30th, 2019)
+
+Maintenance Release + Top-Level API Change
+
+* Added bin/console script
+* Added CHANGELOG
+* Updated README with publication steps, additional testing steps
+* Bumped minimum Ruby to 2.3.8
+* Added Guard
+* Changed Logicality::Logic#evaluate to Logicality#evaluate (breaking change)
